--- src/db/models/base.py
@@ -1,3 +1,0 @@
-from sqlalchemy.ext.declarative import declarative_base
-
-BASE = declarative_base()
