--- src/db/models/performance.py
@@ -1,24 +1,0 @@
-from sqlalchemy import Column, Integer, ForeignKey, String
-from sqlalchemy.orm import relationship
-
-from .base import BASE
-
-
-class Performance(BASE):
-    __tablename__ = 'performance'
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
-
-    player_id = Column(Integer, ForeignKey(f'player.id'))
-
-    aces = Column(Integer)
-    double_faults = Column(Integer)
-    serve_points = Column(Integer)
-    first_serve_in = Column(Integer)
-    first_serve_won = Column(Integer)
-    second_serve_won = Column(Integer)
-    serve_games = Column(Integer)
-    break_points_faced = Column(Integer)
-    break_points_saved = Column(Integer)
-
-    player = relationship("Player", foreign_keys=[player_id])
