+++ tests/db/test_db.py
@@ -3,16 +3,15 @@ import datetime
 import numpy as np
 
 from ..conftest import TEST_DB
-from src.db.db import CommandDB, QueryDB
-from src.db.models.player import Player
-from src.db.models.tournament import Tournament
-from src.db.models.performance import Performance
-from src.db.models.game import WTA, ITF
+from src.db.db import CommandDB, QueryDB, DBClient
+from src.db.models import Player, Tournament, Performance, WTA, ITF
+
+from src.db.schemas import PlayerCreate
 
 
 @pytest.fixture(scope='module')
 def sample_player() -> Player:
-    return Player(first_name='Tom', last_name='Jones', nationality='IRE', dob=datetime.date(2000, 1, 1), hand='R')
+    return Player(first_name='Tom', last_name='Jones', nationality='IRE', dob=datetime.datetime(2000, 1, 1), hand='R')
 
 
 @pytest.fixture(scope='module')
@@ -22,7 +21,7 @@ def sample_tournament() -> Tournament:
 
 class TestDBClient:
 
-    def test_init(self, db_client):
+    def test_init(self, db_client: DBClient):
         '''
         Ensure connected to expected database.
         '''
@@ -36,7 +35,10 @@ class TestCommandDB:
         return CommandDB(db_client.session)
 
     def test_add_player(self, db_client, command_db, sample_player):
-        command_db.add_player(sample_player)
+        player = PlayerCreate(first_name=sample_player.first_name, last_name=sample_player.last_name,
+                              nationality=sample_player.nationality, dob=sample_player.dob, hand=sample_player.hand)
+
+        command_db.add_player(player)
 
         # will raise exception if there is anything other than 1 instance present
         db_client.session.query(Player).\
@@ -58,8 +60,11 @@ class TestCommandDB:
             filter(Tournament.start_date == sample_tournament.start_date).one()
 
     def test_add_game(self, db_client, command_db, sample_player, sample_tournament):
+        # changing names to avoid adding multiple instances of same player
         w_performance = Performance(player=sample_player)
+        w_performance.player.first_name = "Ben"
         l_performance = Performance(player=sample_player)
+        l_performance.player.first_name = "Fran"
 
         game = WTA(tournament=sample_tournament, w_performance=w_performance, l_performance=l_performance)
 
@@ -80,19 +85,22 @@ class TestQueryDB:
     def test_get_player(self, query_db, sample_player):
         player = query_db.get_player(sample_player.name, sample_player.dob)
 
-        assert player == sample_player
+        assert player.name == sample_player.name
         assert query_db.get_player("Test Player", datetime.date.today()) is None
 
     def test_get_tournament(self, query_db, sample_tournament):
         tournament = query_db.get_tournament(sample_tournament.name, sample_tournament.start_date)
 
-        assert tournament == sample_tournament
+        assert tournament.name == sample_tournament.name
         assert query_db.get_tournament("Test", datetime.date.today()) is None
 
     def test_get_game(self, query_db, sample_tournament, sample_player):
-        game = query_db.get_game(sample_tournament, sample_player, sample_player)
+        ben, fran = sample_player, sample_player
+        ben.first_name, fran.first_name = "Ben", "Fran"
+
+        game = query_db.get_game(sample_tournament, ben, fran)
 
         assert game.tournament == sample_tournament
-        assert game.w_performance.player == sample_player
-        assert game.l_performance.player == sample_player
+        assert game.w_performance.player.name == ben.name
+        assert game.l_performance.player.name == fran.name
         assert game.circuit == 'wta'
